Split VectorSet code from ResultProcessor (also: *Lease*)
